requirements.txt
pika
requests
python-dotenv
PyPDF2
Pillow